\docType{methods}
\name{$}
\alias{$}
\alias{$,SummaryUSL-method}
\alias{$,USL-method}
\title{Extract parts of an \code{SummaryUSL} or \code{USL} object}
\arguments{
  \item{x}{Object from which to extract elements.}

  \item{name}{A literal character string or a \link{name}
  (possibly quoted).}
}
\description{
  The operator extracts a part of an object.
}

