\name{finish}
\alias{finish}
\title{Finish USL object after initialize}
\usage{
  finish(.Object)
}
\arguments{
  \item{.Object}{A USL model object returned from
  \code{new}.}
}
\value{
  An object of class USL with updated model parameters.
}
\description{
  \code{finish} updates a USL object and sets additional
  parameters after the object has been initialized.
}
\details{
  This is a package internal function.
}
\seealso{
  \code{\link{usl}}, \code{\link{initialize}}
}
\keyword{internal}

