\docType{methods}
\name{initialize}
\alias{initialize}
\alias{initialize,SummaryUSL-method}
\alias{initialize,USL-method}
\title{Generate an object from the SummaryUSL class}
\arguments{
  \item{.Object}{The SummaryUSL object to initialize.}

  \item{call}{The formula used to create the USL model.}

  \item{coefficients}{A vector containing the coefficients
  of the USL model.}

  \item{.Object}{The SummaryUSL object to initialize.}

  \item{call}{The formula used to create the USL model.}
}
\description{
  Initialize the object.

  Initialize the object.
}

