% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{initialize,USL-method}
\alias{initialize,USL-method}
\title{Generate an object from the "\code{USL}" class}
\usage{
\S4method{initialize}{USL}(.Object, call, frame, regr, resp, alpha, beta, gamma)
}
\arguments{
\item{.Object}{The object to initialize.}

\item{call}{The formula used to create the USL model.}

\item{frame}{The model frame containing the variables in the model.}

\item{regr}{The name of the regressor variable in the model.}

\item{resp}{The name of the response variable in the model.}

\item{alpha}{The contention parameter of the model.}

\item{beta}{The coherency delay parameter of the model.}

\item{gamma}{The slope of the ideal parallel scaling of the three parameter
model. This parameter corrsponds to the scale.factor parameter of the
two parameter model.}
}
\value{
An object of the specific type.
}
\description{
Initialize the object.
}
\keyword{internal}
