\docType{methods}
\name{peak.scalability}
\alias{peak.scalability}
\alias{peak.scalability,USL-method}
\title{Peak scalability value}
\arguments{
  \item{object}{An USL object for the model}
}
\value{
  A numeric value for the point where peak scalability is
  reached.
}
\description{
  Calculate the point of peak scalability for a specific
  model. See formula (4.33) in "Guerilla Capacity Planing".
}

