\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,USL-method}
\title{Plot the result from an Universal Scalability Law model}
\arguments{
  \item{x}{The USL object}

  \item{...}{Other graphical parameters (see
  \code{\link{par}})}
}
\description{
  Draw a line plot for an Universal Scalability Law model.
}

