\docType{methods}
\name{predict}
\alias{predict}
\alias{predict,USL-method}
\title{Universal Scalability Law model predictions}
\arguments{
  \item{object}{An USL model object for which prediction is
  desired.}

  \item{newdata}{An optional data frame in which to look
  for variables with which to predict. If omitted, the
  fitted values are used.}
}
\value{
  An object of class \code{\link{SummaryUSL-class}}
}
\description{
  \code{predict} is a function for predictions from the
  result of the Universal Scalability Law.
}

