\docType{methods}
\name{scalability}
\alias{scalability}
\alias{scalability,USL-method}
\title{Scalability function of the USL model}
\arguments{
  \item{object}{An USL object}
}
\value{
  A function with parameter \code{x} that calculates the
  value of the specific model.
}
\description{
  Return a function to calculate the scalability for a
  specific model.
}

