\docType{methods}
\name{show}
\alias{show}
\alias{show,SummaryUSL-method}
\alias{show,USL-method}
\title{Show objects of class \code{SummaryUSL} or \code{USL}}
\arguments{
  \item{object}{The object to be printed.}
}
\value{
  \code{show} returns an invisible \code{NULL}.
}
\description{
  Display the object by printing it.
}

