% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show,USL-method}
\alias{show,USL-method}
\title{Show objects of class "\code{USL}"}
\usage{
\S4method{show}{USL}(object)
}
\arguments{
\item{object}{The object to be printed.}
}
\value{
\code{show} returns an invisible \code{NULL}.
}
\description{
Display the object by printing it.
}
\examples{
require(usl)

data(raytracer)

## Show USL model
show(usl(throughput ~ processors, raytracer))

}
\seealso{
\code{\link{usl}}, \code{\link{USL-class}}
}
