% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma.R
\name{sigma,USL-method}
\alias{sigma,USL-method}
\title{Extract Residual Standard Deviation 'Sigma'}
\usage{
\S4method{sigma}{USL}(object, ...)
}
\arguments{
\item{object}{An object from class \code{USL}.}

\item{...}{Other arguments passed to other methods.}
}
\value{
A single number.
}
\description{
\code{sigma} Extract Residual Standard Deviation 'Sigma'
}
\examples{
require(usl)

data(raytracer)

## Print result from USL model for demo dataset
print(sigma(usl(throughput ~ processors, raytracer)))

}
\seealso{
\code{\link{usl}}, \code{\link{USL-class}}
}
