\name{SummaryUSL-class}
\alias{SummaryUSL-class}
\title{Class "\code{SummaryUSL}" for summaries of class "\code{USL-class}"}
\description{
  \code{SummaryUSL} is a class to store the summary of an
  \code{\link{USL-class}}.
}
\section{Slots}{
  \describe{ \item{\code{call}:}{The call to calculate the
  model} \item{\code{coefficients}:}{The coefficients sigma
  and kappa of the model} }
}

