\name{gradient.usl}
\alias{gradient.usl}
\title{Calculate gradient for the universal scalability function}
\usage{
gradient.usl(x)
}
\arguments{
  \item{x}{The USL object.}
}
\value{
The gradient matrix.
}
\description{
The implementation of this function has been adopted from
the generated output of the \code{\link{deriv}} function.
}
\seealso{
\code{\link{usl}}
}
\keyword{internal}

