\docType{methods}
\name{print}
\alias{print}
\alias{print,SummaryUSL-method}
\alias{print,USL-method}
\title{Print objects of class \code{SummaryUSL} or \code{USL}}
\arguments{
  \item{x}{An object.}

  \item{digits}{Minimal number of significant digits, see
  \link{print.default}.}

  \item{...}{further arguments passed to other methods.}
}
\value{
  \code{print} returns the object \code{x} invisibly.
}
\description{
  \code{print} prints its argument and returns is invisibly
  (via \code{\link{invisible}(x)}).
}

