\docType{class}
\name{SummaryUSL-class}
\alias{SummaryUSL-class}
\title{Class "\code{SummaryUSL}" for summaries of class "\code{USL-class}"}
\description{
  \code{SummaryUSL} is a class to store the summary of an
  \code{\link{USL-class}}.
}
\section{Slots}{
  \describe{ \item{\code{call}:}{The call used to create
  the model.} \item{\code{residuals}:}{The residuals of the
  model. This is a vector.}
  \item{\code{r.squared}:}{Coefficient of determination of
  the model.} \item{\code{adj.r.squared}:}{Adjusted
  coefficient of determination.}
  \item{\code{coefficients}:}{The coefficients sigma and
  kappa of the model.} }
}

