\name{USL-class}
\alias{USL-class}
\title{Class "\code{USL}" for Universal Scalability Law models}
\description{
  This class encapsulates the Universal Scalability Law.
  Use the function \code{\link{usl}} to create new objects
  from this class.
}
\details{
  The class contains the class
  "\code{\link{SummaryUSL-class}}".
}
\section{Slots}{
  \describe{ \item{\code{frame}:}{The model frame.}
  \item{\code{regr}:}{The name of the regressor variable.}
  \item{\code{regr}:}{The name of the response variable.}
  \item{\code{scale.factor}:}{The scale factor of the
  model.} \item{\code{fitted.values}:}{The fitted values of
  the model. This is a vector.} }
}
\references{
  N. J. Gunther. Guerrilla Capacity Planning.
  Springer-Verlag, Heidelberg, Germany, 2007.
}
\seealso{
  \code{\link{SummaryUSL-class}}, \code{\link{usl}}
}

