% Generated by roxygen2 (4.0.2): do not edit by hand
\name{usl.solve.nls}
\alias{usl.solve.nls}
\title{Solve a USL model using non linear regression}
\usage{
usl.solve.nls(model)
}
\arguments{
\item{model}{A data frame with two columns containing the values of the
  predictor variable in the first column and the values of the response
  variable in the second column.}
}
\value{
A list containing three elements: the scale.factor of the model,
  the model coefficients sigma and kappa.
}
\description{
This function solves a USL model using non linear regression with least
squares. It uses the function \code{\link{nls}} with the "\code{port}"
algorithm to perform the calculation. All restrictions of the algorithm
apply.
}
\seealso{
\code{\link{usl}}
}
\keyword{internal}

