% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usl.R
\name{usl.solve.nlxb}
\alias{usl.solve.nlxb}
\title{Solve a USL model using non linear regression}
\usage{
usl.solve.nlxb(model)
}
\arguments{
\item{model}{A data frame with two columns containing the values of the
predictor variable in the first column and the values of the response
variable in the second column.}
}
\value{
A list containing three elements: the scale.factor of the model,
  the model coefficients alpha and beta.
}
\description{
This function solves a USL model using non linear regression with least
squares. It uses the function \code{\link{nlxb}} from the \pkg{nlsr}
package to perform the calculation.
}
\references{
John C. Nash. nlsr: Functions for nonlinear least squares
  solutions, 2017. R package version 2017.6.18.
}
\seealso{
\code{\link{usl}}
}
\keyword{internal}
